Update calls for support for newer plugin-sdk and VS2022
Not all calls are supported, a couple of preferences don't have related
settings on newer versions of plugin-sdk (or I might just be looking at
the wrong direction), so some options in the settings menu might not work.

+++ source/Utility.h
@@ -24,7 +24,7 @@
 #define clamp(v, low, high) ((v)<(low) ? (low) : (v)>(high) ? (high) : (v))
 #define ARRAY_SIZE(array) (sizeof(array) / sizeof(array[0]))
 
-#define FLASH_ITEM(on, off) (CTimer::m_snTimeInMillisecondsPauseMode % on + off < on)
+#define FLASH_ITEM(on, off) (static_cast<int>(CTimer::m_snTimeInMillisecondsPauseMode) % on + off < on)
 
 extern bool hudSafeZones;
 extern void SetHUDSafeZone(bool on);
